PAL3A gameplay WIP
* Implement more PAL3A specific commands.
* Fix data structure issues with PAL3A ScnObjectInfo.
* And more fixes/tweaks for supporting PAL3A game play.

+++ Assets/Scripts/Pal3/Scene/SceneObjects/ClimbableObject.cs
@@ -38,9 +38,9 @@ namespace Pal3.Scene.SceneObjects
                 // Parameters[4] == 1 ? Crossing different layer : Same layer
 
                 CommandDispatcher<ICommand>.Instance.Dispatch(new PlayerActorClimbObjectCommand(Info.Id,
-                    new Vector2Int(Info.Parameters[0], Info.Parameters[1]),
-                    new Vector2Int(Info.Parameters[2], Info.Parameters[3]),
-                    Info.Parameters[4] == 1));
+                    new Vector2Int((int)Info.Parameters[0], (int)Info.Parameters[1]),
+                    new Vector2Int((int)Info.Parameters[2], (int)Info.Parameters[3]),
+                    (int)Info.Parameters[4] == 1));
             }
             else
             {
